kicker motor dribbler updates
