+++ kicker-board/src/bin/hwtest-coms/main.rs
@@ -29,7 +29,7 @@ use ateam_kicker_board::{
     pins::*, tasks::{get_system_config, ClkSource}
 };
 
-use ateam_lib_stm32::{idle_buffered_uart_spawn_tasks, static_idle_buffered_uart, static_idle_buffered_uart_nl, uart::queue::{UartReadQueue, UartWriteQueue}};
+use ateam_lib_stm32::{idle_buffered_uart_spawn_tasks, static_idle_buffered_uart_nl, uart::queue::{UartReadQueue, UartWriteQueue}};
 
 use ateam_common_packets::bindings::{KickerControl, KickerTelemetry, KickRequest};
 
@@ -47,6 +47,7 @@ fn get_empty_control_packet() -> KickerControl {
         _bitfield_1: KickerControl::new_bitfield_1(0, 0, 0),
         kick_request: KickRequest::KR_DISABLE,
         kick_speed: 0.0,
+        drib_speed: 0.0,
     }
 }
